+++ client/cache/error_code.go
@@ -124,7 +124,12 @@ SSDB cache only accept string value. Please check your input.
 `)
 
 var InvalidLoadFunc = berror.DefineCode(4002023, moduleName, "InvalidLoadFunc", `
-Invalid loadFunc.
+Invalid load function for read-through pattern decorator.
+You should pass a valid(non-nil) load function when initiate the decorator instance.
+`)
+
+var LoadFuncFailed = berror.DefineCode(4002024, moduleName, "InvalidLoadFunc", `
+Failed to load data, please check whether the loadfunc is correct
 `)
 
 var DeleteFileCacheItemFailed = berror.DefineCode(5002001, moduleName, "DeleteFileCacheItemFailed", `