+++ client/cache/singleflight.go
@@ -1,11 +1,26 @@
+// Copyright 2014 beego Author. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package cache
 
 import (
 	"context"
 	"fmt"
+	"time"
+
 	"github.com/beego/beego/v2/core/berror"
 	"golang.org/x/sync/singleflight"
-	"time"
 )
 
 // SingleflightCache
@@ -19,7 +34,8 @@ type SingleflightCache struct {
 
 // NewSingleflightCache create SingleflightCache
 func NewSingleflightCache(c Cache, expiration time.Duration,
-	loadFunc func(ctx context.Context, key string) (any, error)) (Cache, error) {
+	loadFunc func(ctx context.Context, key string) (any, error),
+) (Cache, error) {
 	if loadFunc == nil {
 		return nil, berror.Error(InvalidLoadFunc, "loadFunc cannot be nil")
 	}