+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # developing
 
+# v2.0.7
+- [Upgrade github.com/go-kit/kit, CVE-2022-24450](https://github.com/beego/beego/pull/5121)
+# v2.0.6
+- [fix: revise the body wrapper to handle empty body case](https://github.com/beego/beego/pull/5102)
+
 # v2.0.5
 
 Note: now we force the web admin service serving HTTP only.
@@ -10,6 +15,7 @@ Note: now we force the web admin service serving HTTP only.
 - [Fix 5012: fix some bug, pass []any as any in variadic function](https://github.com/beego/beego/pull/5012)
 - [Fix 5022: Miss assigning listener to graceful Server](https://github.com/beego/beego/pull/5028)
 - [Fix 4955: Make commands and Docker compose for ORM unit tests](https://github.com/beego/beego/pull/5031)
+- [add read through for cache module](https://github.com/beego/beego/pull/5116)
 
 
 # v2.0.4